requirements.txt
Cython==0.29.21
cython-bbox==0.1.3
dcor==0.5.2
decorator==4.4.2
decord==0.6.0
detectron2==0.5+cu101
dlib==19.22.99
easydict==1.9
gdown==3.12.2
gluoncv==0.11.0b20210719
h5py==3.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
imutils==0.5.4
matplotlib==3.3.2
mxnet-cu101mkl
numba==0.51.2
numpy==1.19.0
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-contrib-python-headless==4.4.0.46
opencv-python==4.2.0.34
opencv-python-headless==4.4.0.46
pandas==1.1.3
Pillow==7.2.0
python-utils==2.5.6
pytorch-lightning==1.3.3
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.22
scikits.bootstrap==1.0.1
scipy==1.5.2
seaborn==0.11.0
sklearn==0.0
torch==1.8.1
torchmetrics==0.4.1
torchsummary==1.5.1
torchvision==0.9.1
tqdm==4.50.0
urllib3==1.26.4
